Change shell script to ruby
This commit also adds a curl call to find out how long a particular
artifact from a set of commits sits in aptly for.

The curl doesn't really work with metadata and this commit is more for
backup, since I will use a new way to determine the duration it sits in
aptly in the next commit.